Rename to avoid conflicts between function in __init__ and module
