+++ tests/test_time_cost.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import pytest
 
-from src.imgdataconvertcodegen.code_exec.time_cost import time_cost, time_cost_in_kg
+from src.imgdataconvertcodegen.code_exec.time_cost_measure import time_cost, time_cost_in_kg
 from src.imgdataconvertcodegen.knowledge_graph_construction import KnowledgeGraph, encode_metadata
 from .data_for_tests.nodes_edges import test_edges
 
@@ -41,7 +41,7 @@ def test_time_cost_to_throw_exception(kg):
 
 def test_time_cost_in_kg(kg):
     costs = [1, 2, 3, 4]
-    with patch('src.imgdataconvertcodegen.code_exec.time_cost.time_cost') as mock_time_cost:
+    with patch('src.imgdataconvertcodegen.code_exec.time_cost_measure.time_cost') as mock_time_cost:
         mock_time_cost.side_effect = costs
 
         actual = time_cost_in_kg(kg, test_img_size=(256, 256), repeat_count=10)
